requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.3.0
backoff==2.2.1
banks==2.2.0
beautifulsoup4==4.14.2
cachetools==6.2.0
cbor==1.0.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
effdet==0.4.1
emoji==2.15.0
fastapi==0.118.0
filelock==3.19.1
filetype==1.2.0
FlagEmbedding==1.2.10
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
google-api-core==2.25.1
google-auth==2.41.1
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ijson==3.4.0
inscriptis==2.6.0
ir_datasets==0.5.11
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
kiwisolver==1.4.9
langdetect==1.0.9
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.3
llama-index-cli==0.5.1
llama-index-core==0.14.3
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-openai==0.5.6
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.5.0
llama-parse==0.6.54
lxml==6.0.2
lz4==4.4.4
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.8.4
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.2.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
olefile==0.47
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.23.0
openai==1.109.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250506
peft==0.17.1
pi_heif==1.1.1
pikepdf==9.11.0
pillow==11.3.0
platformdirs==4.4.0
portalocker==3.2.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyMuPDF==1.26.4
pyparsing==3.2.5
pypdf==6.1.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
timm==1.0.20
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.2
trec-car-tools==2.6
triton==3.4.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unlzw3==0.2.3
unstructured==0.18.15
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.30.3
uvloop==0.21.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xlsxwriter==3.2.9
xxhash==3.5.0
yarl==1.20.1
zlib-state==0.1.10
pydantic-settings==2.6.1
